Move pure js funtions into own modules
- Moved functions out of Game component and into own module
- https://chicagotechgrowhouse.atlassian.net/browse/MKA-89
